chore: split tabs component
